+++ tests/conftest.py
@@ -150,6 +150,10 @@ def emr():
     return make_client("emr")
 
 @pytest.fixture(scope="session")
+def elbv2():
+    return make_client("elbv2")
+                                                                                                          
+@pytest.fixture(scope="session")
 def ebs():
     return make_client("ec2")